Refactor/backend core

+++ src/main/java/com/lingulu/service/DashboardService.java
@@ -31,7 +31,7 @@ public class DashboardService {
         CourseProgress courseProgress = courseProgressRepository.findActiveCourse(userId, ProgressStatus.IN_PROGRESS)
                                         .orElseGet(() -> courseProgressRepository.findByUser_UserIdAndCourse_Position(userId, courseProgressRepository.countByUser_UserId(userId)));
         int streak = userLearningStatsRepository.getStreak(userId);
-        float progressPercentage = courseProgress.getCompletedLessons() * 100 / courseProgress.getTotalLessons();
+        float progressPercentage = (float) (courseProgress.getCompletedLessons() * 100) / courseProgress.getTotalLessons();
         CourseResponse courseResponse = CourseResponse.builder()
                                         .courseId(courseProgress.getCourse().getCourseId())
                                         .courseTitle(courseProgress.getCourse().getCourseTitle())
